feat ✨: added calculator features
