+++ src/cmd/version.go
@@ -9,14 +9,10 @@ import (
 // versionCmd represents the version command
 var versionCmd = &cobra.Command{
 	Use:   "version",
-	Short: "A brief description of your command",
-	Long: `A longer description that spans multiple lines and likely contains examples
-and usage of using your command. For example:
-
-Cobra is a CLI library for Go that empowers applications.
-This application is a tool to generate the needed files
-to quickly create a Cobra application.`,
+	Short: "Display application version information.",
+	Long: `The version command provides information about the application's version.
+Use this command to check the current version of the application.`,
 	Run: func(cmd *cobra.Command, args []string) {
-		fmt.Println("Azureprice CLI version 0.1.0")
+		fmt.Println("Azureprice CLI version 0.0.1")
 	},
 }
